requirements.txt
numpy==1.26.4
pydantic==2.10.6
tiktoken==0.7.0
openai==1.61.1
langchain==0.3.19
langchain_community==0.3.18
langchain-openai==0.3.6
faiss-cpu==1.8.0
tqdm
requests
python-dotenv
